refactor(gc): split roots/scan_mode.rs out; delete the host-pressure conservative arm outright
Two follow-ups the compiler found.

1. gc/roots.rs crossed the 2000-line file-size gate (1992 -> 2035). The
   conservative-scan mode block is topically self-contained -- the mode enum,
   its env/override precedence, ManualGcScanGuard, and the decision the mark
   phase consults -- so it moves to roots/scan_mode.rs. roots.rs is now 1851.

2. ConservativeScanSite::HostPressure was never constructed, because after the
   deferral js_gc_memory_pressure has NO conservative arm left: it either
   collects with precise roots or defers. So the variant is deleted rather than
   kept 'for completeness'. An enum arm nothing can produce is a claim no test
   can check -- the host-pressure tests now assert automatic_scan_fallback_total
   == 0 instead, which catches reintroduction at any automatic site.

The counter read APIs are #[cfg(test)]; ops observability is the PERRY_GC_DIAG
line, which is unconditional.
